edición para tener datos separados
